+++ lib/globals.ts
@@ -7,11 +7,18 @@
  *
  * @module globals
  */
-"use strict";
 
+import {configure} from "react-hotkeys";
 import "util.obj-cycle";
 import "util.string";
 import {globalize} from "./shared/helpers";
 
 globalize("hljs", require("highlight.js"));
 globalize("Quill", require("quill"));
+
+// Configuration options for react HotKeys.  This turns off the tag
+// ignore function.  If this is not turned off, then the default
+// key handler will not "bubble up" events in composite controls.
+configure({
+	ignoreTags: []
+});